Remove SharedConstant.FieldTypes and int cast #2 #38

+++ Gravity/Gravity.Test/TestClasses/GravityLevel2Child.cs
@@ -8,10 +8,10 @@ namespace Gravity.Test.TestClasses
 	public class GravityLevel2Child : BaseDto
 	{
 		[RelativityObjectFieldParentArtifactId]
-		[RelativityObjectField("2EB737C1-564B-4E2B-9C0A-CB34F3FB998C", (int)RdoFieldType.SingleObject, typeof(GravityLevelOne))]
+		[RelativityObjectField("2EB737C1-564B-4E2B-9C0A-CB34F3FB998C", RdoFieldType.SingleObject, typeof(GravityLevelOne))]
 		public int GavityLevelOneArtifactId { get; set; }
 
-		[RelativityObjectField("DAD68D2A-F740-4473-8047-E89C5C58D987", (int)RdoFieldType.FixedLengthText, 255)]
+		[RelativityObjectField("DAD68D2A-F740-4473-8047-E89C5C58D987", RdoFieldType.FixedLengthText, 255)]
 		public override string Name { get; set; }
 	}
 }
